feature/ASAA-25 cleanup on GoogleMapsView moving Markers to it GoogleMapsComponents
